Move shuffle fn from data access to model
